// src/types/brush/enum.ts
/**
 * @enum {string} list of brush shapes.
 * @prop {string} square - square figure to paint on the canvas
 * @prop {string} rectangle - rectangle figure to paint on the canvas
 * @prop {string} line - line figure to paint on the canvas
 */
export enum ShapesBrush {
  square = 'square',
  rectangle = 'rectangle',
  line = 'line',
}